+++ src/common/data_structs/src/array.c
@@ -1,5 +1,4 @@
-#include <stdlib.h>
-#include <string.h>
+#include <string.h>  /* memcpy */
 
 #include "common/lang/assert.h"
 #include "common/lang/mem.h"
@@ -7,10 +6,10 @@
 #include "common/data_structs/array.h"
 #include "common/data_structs/arrayrep.h"
 
-void ArrayRep_init(Array_T array, size_t length, size_t size, void *mem_loc) {
+void ArrayRep_init(Array_T array, size_t length, size_t size, void* mem_loc)
+{
   Assert(array);
   Assert((mem_loc && length > 0) || (length == 0 && mem_loc == NULL));
-
   Assert(size > 0);
 
   array->length = length;
@@ -22,7 +21,8 @@ void ArrayRep_init(Array_T array, size_t length, size_t size, void *mem_loc) {
     array->mem_array = NULL;
 }
 
-Array_T Array_new(int length, size_t size) {
+Array_T Array_new(size_t length, size_t size)
+{
   Array_T array;
   NEW(array);
 
@@ -34,31 +34,36 @@ Array_T Array_new(int length, size_t size) {
   return array;
 }
 
-void Array_free(Array_T *array) {
+void Array_free(Array_T* array)
+{
   Assert(array && *array);
 
   FREE((*array)->mem_array);
   FREE(*array);
 }
 
-size_t Array_length(Array_T array) {
+size_t Array_length(Array_T array)
+{
   Assert(array);
   return array->length;
 }
 
-size_t Array_size(Array_T array) {
+size_t Array_size(Array_T array)
+{
   Assert(array);
   return array->size;
 }
 
-void *Array_get(Array_T array, unsigned i) {
+void* Array_get(Array_T array, unsigned i)
+{
   Assert(array);
   Assert(i < array->length);
 
   return array->mem_array + i*array->size;
 }
 
-void *Array_put(Array_T array, unsigned i, void *elem) {
+void* Array_put(Array_T array, unsigned i, void* elem)
+{
   Assert(array);
   Assert(i < array->length);
   Assert(elem);
@@ -68,7 +73,8 @@ void *Array_put(Array_T array, unsigned i, void *elem) {
   return elem;
 }
 
-void Array_resize(Array_T array, size_t length) {
+void Array_resize(Array_T array, size_t length)
+{
   Assert(array);
 
   if (length == 0)
@@ -82,7 +88,8 @@ void Array_resize(Array_T array, size_t length) {
   array->length = length;
 }
 
-Array_T Array_copy(Array_T array, size_t length) {
+Array_T Array_copy(Array_T array, size_t length)
+{
   Array_T copy;
 
   Assert(array);
@@ -90,10 +97,10 @@ Array_T Array_copy(Array_T array, size_t length) {
   copy = Array_new(length, array->size);
 
   if (copy->length >= array->length && array->length > 0)
-    memcpy(copy->mem_array, array->mem_array, array->length*array->size);
+    memcpy(copy->mem_array, array->mem_array, array->length * array->size);
 
   else if (array->length > copy->length && copy->length > 0)
-    memcpy(copy->mem_array, array->mem_array, copy->length*array->size);
+    memcpy(copy->mem_array, array->mem_array, copy->length * array->size);
 
   return copy;
 }
